add watcher mode and release 0.1.0

+++ lib/prss/downloader.rb
@@ -10,13 +10,18 @@ module PRSS
       @hydra = Typhoeus::Hydra.new
     end
 
-    def download_to(output_path)
+    def self.verify!(output_path)
       output = Pathname.new(output_path).expand_path
       raise 'output path is not directory' unless output.directory?
+      output
+    end
+
+    def download_to(output_path)
+      output = self.class.verify!(output_path)
 
       @files = []
       @links.each do |link|
-        request = Typhoeus::Request.new(link, :follow_location => true)
+        request = Typhoeus::Request.new(link, followlocation: true)
 
         request.on_complete do |response|
           @files << save_file(response, output)
@@ -30,9 +35,11 @@ module PRSS
       @files
     end
 
-    def save_file response, output
+    def save_file(response, output)
       filename = response.headers_hash['Content-Disposition'][/filename="(.+)"$/ ,1]
       file = output.join(filename)
+
+      return if file.exist?
 
       open(file, 'wb') do |file|
         file.write(response.body)
